Fix mod for update 1.1
For some reason `SetActorsAI 0` (when run on a ref in another cell) seems to immediately CTD in the new 1.1 update. So I am commenting that out in all the result scripts for now. It doesn't seem to affect the collision detection at all, so I think it is fine to leave commented until hopefully virtuous fixes it.